Simplify graphics api

+++ src/graphics/graphics.h
@@ -6,57 +6,17 @@
 #include "../math/quat.h"
 #include "../math/vec.h"
 #include "camera.h"
-#include "camera_instance.h"
 #include "material.h"
 #include "mesh.h"
 #include "render_target.h"
 #include "rgb_spectrum.h"
 #include "scene.h"
-#include "scene_diff.h"
-#include "scene_node.h"
 #include "surface.h"
-#include "surface_instance.h"
 #include "texture.h"
 
 namespace marlon {
 namespace graphics {
-class Camera;
-struct Camera_create_info;
-class Camera_instance;
-struct Camera_instance_create_info;
-class Render_target;
-class Render_stream;
-struct Render_stream_create_info;
-class Scene;
-struct Scene_create_info;
-class Scene_diff;
-struct Scene_diff_create_info;
-class Scene_node;
-struct Scene_node_create_info;
-class Surface_instance;
-struct Surface_instance_create_info;
-
 class Graphics;
-
-class Texture_deleter {
-public:
-  Texture_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
-
-  void operator()(Texture *texture) const noexcept;
-
-private:
-  Graphics *_owner;
-};
-
-// class Material_deleter {
-// public:
-//   Material_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
-
-//   void operator()(Material *material) const noexcept;
-
-// private:
-//   Graphics *_owner;
-// };
 
 class Mesh_deleter {
 public:
@@ -68,15 +28,25 @@ private:
   Graphics *_owner;
 };
 
-// class Surface_deleter {
-// public:
-//   Surface_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
+class Texture_deleter {
+public:
+  Texture_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
 
-//   void operator()(Surface *surface) const noexcept;
+  void operator()(Texture *texture) const noexcept;
 
-// private:
-//   Graphics *_owner;
-// };
+private:
+  Graphics *_owner;
+};
+
+class Material_deleter {
+public:
+  Material_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
+
+  void operator()(Material *material) const noexcept;
+
+private:
+  Graphics *_owner;
+};
 
 class Scene_deleter {
 public:
@@ -88,26 +58,44 @@ private:
   Graphics *_owner;
 };
 
-class Scene_diff_deleter {
+class Surface_deleter {
 public:
-  Scene_diff_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
+  Surface_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
 
-  void operator()(Scene_diff *scene_diff) const noexcept;
+  void operator()(Surface *surface) const noexcept;
 
 private:
   Graphics *_owner;
 };
 
-using Unique_texture_ptr = std::unique_ptr<Texture, Texture_deleter>;
-// using Unique_material_ptr = std::unique_ptr<Material, Material_deleter>;
+class Camera_deleter {
+public:
+  Camera_deleter(Graphics *owner = nullptr) noexcept : _owner{owner} {}
+
+  void operator()(Camera *camera) const noexcept;
+
+private:
+  Graphics *_owner;
+};
+
 using Unique_mesh_ptr = std::unique_ptr<Mesh, Mesh_deleter>;
-// using Unique_surface_ptr = std::unique_ptr<Surface, Surface_deleter>;
+using Unique_texture_ptr = std::unique_ptr<Texture, Texture_deleter>;
+using Unique_material_ptr = std::unique_ptr<Material, Material_deleter>;
 using Unique_scene_ptr = std::unique_ptr<Scene, Scene_deleter>;
-using Unique_scene_diff_ptr = std::unique_ptr<Scene_diff, Scene_diff_deleter>;
+using Unique_surface_ptr = std::unique_ptr<Surface, Surface_deleter>;
+using Unique_camera_ptr = std::unique_ptr<Camera, Camera_deleter>;
 
 class Graphics {
 public:
   virtual ~Graphics() = default;
+
+  virtual Mesh *create_mesh(Mesh_create_info const &create_info) = 0;
+
+  virtual void destroy_mesh(Mesh *mesh) noexcept = 0;
+
+  Unique_mesh_ptr create_mesh_unique(Mesh_create_info const &create_info) {
+    return Unique_mesh_ptr{create_mesh(create_info), this};
+  }
 
   virtual Texture *create_texture(Texture_create_info const &create_info) = 0;
 
@@ -118,32 +106,15 @@ public:
     return Unique_texture_ptr{create_texture(create_info), this};
   }
 
-  // virtual Material *
-  // create_material(Material_create_info const &create_info) = 0;
+  virtual Material *
+  create_material(Material_create_info const &create_info) = 0;
 
-  // virtual void destroy_material(Material *material) noexcept = 0;
+  virtual void destroy_material(Material *material) noexcept = 0;
 
-  // Unique_material_ptr
-  // create_material_unique(Material_create_info const &create_info) {
-  //   return Unique_material_ptr{create_material(create_info), this};
-  // }
-
-  virtual Mesh *create_mesh(Mesh_create_info const &create_info) = 0;
-
-  virtual void destroy_mesh(Mesh *mesh) noexcept = 0;
-
-  Unique_mesh_ptr create_mesh_unique(Mesh_create_info const &create_info) {
-    return Unique_mesh_ptr{create_mesh(create_info), this};
+  Unique_material_ptr
+  create_material_unique(Material_create_info const &create_info) {
+    return Unique_material_ptr{create_material(create_info), this};
   }
-
-  // virtual Surface *create_surface(Surface_create_info const &create_info) = 0;
-
-  // virtual void destroy_surface(Surface *surface) noexcept = 0;
-
-  // Unique_surface_ptr
-  // create_surface_unique(Surface_create_info const &create_info) {
-  //   return Unique_surface_ptr{create_surface(create_info), this};
-  // }
 
   virtual Scene *create_scene(Scene_create_info const &create_info) = 0;
 
@@ -153,55 +124,69 @@ public:
     return Unique_scene_ptr{create_scene(create_info), this};
   }
 
-  virtual Scene_diff *
-  create_scene_diff(Scene_diff_create_info const &create_info) = 0;
+  virtual Surface *create_surface(Surface_create_info const &create_info) = 0;
 
-  virtual void destroy_scene_diff(Scene_diff *scene_diff) noexcept = 0;
+  virtual void destroy_surface(Surface *surface) noexcept = 0;
 
-  Unique_scene_diff_ptr
-  create_scene_diff_unique(Scene_diff_create_info const &create_info) {
-    return Unique_scene_diff_ptr{create_scene_diff(create_info), this};
+  Unique_surface_ptr
+  create_surface_unique(Surface_create_info const &create_info) {
+    return Unique_surface_ptr{create_surface(create_info), this};
   }
 
-  virtual void apply_scene_diff(Scene_diff *scene_diff, float factor) = 0;
+  virtual Camera *create_camera(Camera_create_info const &create_info) = 0;
 
-  virtual void apply_scene_diff(Scene_diff *scene_diff) = 0;
+  virtual void destroy_camera(Camera *camera) noexcept = 0;
+
+  Unique_camera_ptr
+  create_camera_unique(Camera_create_info const &create_info) {
+    return Unique_camera_ptr{create_camera(create_info), this};
+  }
 
   // Render_target creation is implementation-specific
 
   virtual void destroy_render_target(Render_target *target) noexcept = 0;
 
-  virtual void render(Scene *source_scene,
-                      Camera_instance *source_camera_instance,
+  virtual void render(Scene *source_scene, Camera *source_camera,
                       Render_target *target) = 0;
 
   // TODO: interface for compositing or figure out how multi view rendering will
   // work compositing could be used for first person view models
 };
 
-inline void Texture_deleter::operator()(Texture *texture) const noexcept {
-  _owner->destroy_texture(texture);
-}
-
-// inline void Material_deleter::operator()(Material *material) const noexcept {
-//   _owner->destroy_material(material);
-// }
-
 inline void Mesh_deleter::operator()(Mesh *mesh) const noexcept {
-  _owner->destroy_mesh(mesh);
+  if (_owner) {
+    _owner->destroy_mesh(mesh);
+  }
 }
 
-// inline void Surface_deleter::operator()(Surface *surface) const noexcept {
-//   _owner->destroy_surface(surface);
-// }
+inline void Texture_deleter::operator()(Texture *texture) const noexcept {
+  if (_owner) {
+    _owner->destroy_texture(texture);
+  }
+}
+
+inline void Material_deleter::operator()(Material *material) const noexcept {
+  if (_owner) {
+    _owner->destroy_material(material);
+  }
+}
+
+inline void Surface_deleter::operator()(Surface *surface) const noexcept {
+  if (_owner) {
+    _owner->destroy_surface(surface);
+  }
+}
 
 inline void Scene_deleter::operator()(Scene *scene) const noexcept {
-  _owner->destroy_scene(scene);
+  if (_owner) {
+    _owner->destroy_scene(scene);
+  }
 }
 
-inline void
-Scene_diff_deleter::operator()(Scene_diff *scene_diff) const noexcept {
-  _owner->destroy_scene_diff(scene_diff);
+inline void Camera_deleter::operator()(Camera *camera) const noexcept {
+  if (_owner) {
+    _owner->destroy_camera(camera);
+  }
 }
 } // namespace graphics
 } // namespace marlon
